# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShiroScans API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface MangaItem {
  id: string;
  title: string;
  coverImage: string;
  provider: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  latestChapter?: string | null;
  genres: string[];
  isNew: boolean;
  /** @nullable */
  updatedAt?: string | null;
}

export interface MangaPage {
  items: MangaItem[];
  page: number;
  hasMore: boolean;
}

export interface Chapter {
  id: string;
  number: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  releasedAt?: string | null;
  /** @nullable */
  views?: number | null;
}

export interface MangaSeries {
  id: string;
  title: string;
  coverImage: string;
  /** @nullable */
  bannerImage?: string | null;
  provider: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  description?: string | null;
  genres: string[];
  /** @nullable */
  author?: string | null;
  /** @nullable */
  artist?: string | null;
  chapters: Chapter[];
  totalChapters: number;
  alternativeTitles: string[];
  /** @nullable */
  serialization?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ChapterPages {
  seriesId: string;
  chapterId: string;
  provider: string;
  pages: string[];
  currentChapter: string;
  /** @nullable */
  prevChapterId?: string | null;
  /** @nullable */
  nextChapterId?: string | null;
  seriesTitle: string;
  /** @nullable */
  chapterTitle?: string | null;
}

export interface HomeFeed {
  featured: MangaItem[];
  latestUpdates: MangaItem[];
  popularNow: MangaItem[];
  newSeries: MangaItem[];
}

export interface Tag {
  id: string;
  name: string;
  /** @nullable */
  count?: number | null;
}

export interface AuthUser {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  role: string;
  createdAt: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface BookmarkItem {
  id: number;
  userId: number;
  provider: string;
  seriesId: string;
  title: string;
  coverImage: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  latestChapter?: string | null;
  createdAt: string;
}

export interface BookmarkInput {
  provider: string;
  seriesId: string;
  title: string;
  coverImage: string;
  type?: string;
  status?: string;
  latestChapter?: string;
}

export interface HistoryItem {
  id: number;
  userId: number;
  provider: string;
  seriesId: string;
  chapterId: string;
  title: string;
  coverImage: string;
  chapterNumber: string;
  readAt: string;
}

export interface HistoryInput {
  provider: string;
  seriesId: string;
  chapterId: string;
  title: string;
  coverImage: string;
  chapterNumber: string;
}

export interface ReactionSummary {
  love: number;
  fire: number;
  wow: number;
  sad: number;
  angry: number;
  /** @nullable */
  userReaction?: string | null;
}

export type ReactionInputReaction = typeof ReactionInputReaction[keyof typeof ReactionInputReaction];


export const ReactionInputReaction = {
  love: 'love',
  fire: 'fire',
  wow: 'wow',
  sad: 'sad',
  angry: 'angry',
} as const;

export interface ReactionInput {
  reaction: ReactionInputReaction;
}

export interface FavouriteItem {
  id: number;
  userId: number;
  provider: string;
  seriesId: string;
  title: string;
  coverImage: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  status?: string | null;
  createdAt: string;
}

export interface FavouriteInput {
  provider: string;
  seriesId: string;
  title: string;
  coverImage: string;
  type?: string;
  status?: string;
}

export interface UserStats {
  bookmarksCount: number;
  favouritesCount: number;
  chaptersRead: number;
}

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  role: string;
  createdAt: string;
  stats: UserStats;
}

export interface ProfileUpdate {
  username?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface PasswordChange {
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export type GetLatestMangaParams = {
page?: number;
provider?: string;
type?: string;
status?: string;
genre?: string;
};

export type GetPopularMangaParams = {
page?: number;
provider?: string;
type?: string;
status?: string;
genre?: string;
};

export type SearchMangaParams = {
q: string;
page?: number;
provider?: string;
};

export type GetRandomMangaParams = {
provider?: string;
};