# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShiroScans API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /manga/latest:
    get:
      operationId: getLatestManga
      tags: [manga]
      summary: Get latest manga updates
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: false
        - in: query
          name: provider
          schema:
            type: string
          required: false
        - in: query
          name: type
          schema:
            type: string
          required: false
        - in: query
          name: status
          schema:
            type: string
          required: false
        - in: query
          name: genre
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Latest manga list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaPage"
  /manga/popular:
    get:
      operationId: getPopularManga
      tags: [manga]
      summary: Get popular manga
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: false
        - in: query
          name: provider
          schema:
            type: string
          required: false
        - in: query
          name: type
          schema:
            type: string
          required: false
        - in: query
          name: status
          schema:
            type: string
          required: false
        - in: query
          name: genre
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Popular manga list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaPage"
  /manga/search:
    get:
      operationId: searchManga
      tags: [manga]
      summary: Search manga
      parameters:
        - in: query
          name: q
          schema:
            type: string
          required: true
        - in: query
          name: page
          schema:
            type: integer
          required: false
        - in: query
          name: provider
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaPage"
  /manga/series/{provider}/{id}:
    get:
      operationId: getMangaSeries
      tags: [manga]
      summary: Get manga series details
      parameters:
        - in: path
          name: provider
          schema:
            type: string
          required: true
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Manga series details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaSeries"
  /manga/chapters/{provider}/{seriesId}/{chapterId}:
    get:
      operationId: getChapterPages
      tags: [manga]
      summary: Get chapter page images
      parameters:
        - in: path
          name: provider
          schema:
            type: string
          required: true
        - in: path
          name: seriesId
          schema:
            type: string
          required: true
        - in: path
          name: chapterId
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Chapter pages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterPages"
  /manga/random:
    get:
      operationId: getRandomManga
      tags: [manga]
      summary: Get a random manga
      parameters:
        - in: query
          name: provider
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Random manga
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaItem"
...
  /favourites:
    get:
      operationId: getFavourites
      tags: [favourites]
      summary: Get user favourites
      responses:
        "200":
          description: Favourites list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FavouriteItem"
    post:
      operationId: addFavourite
      tags: [favourites]
      summary: Add favourite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavouriteInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavouriteItem"
  /favourites/{provider}/{seriesId}:
    delete:
      operationId: removeFavourite
      tags: [favourites]
      summary: Remove favourite
      parameters:
        - in: path
          name: provider
          schema:
            type: string
          required: true
        - in: path
          name: seriesId
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    MangaItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        coverImage:
          type: string
        provider:
          type: string
        type:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        latestChapter:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        isNew:
          type: boolean
        updatedAt:
          type: ["string", "null"]
      required:
        - id
        - title
        - coverImage
        - provider
        - isNew
        - genres
    MangaPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MangaItem"
        page:
          type: integer
        hasMore:
          type: boolean
      required:
        - items
        - page
        - hasMore
    Chapter:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        title:
          type: ["string", "null"]
        releasedAt:
          type: ["string", "null"]
        views:
          type: ["integer", "null"]
      required:
        - id
        - number
    MangaSeries:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        coverImage:
          type: string
        bannerImage:
          type: ["string", "null"]
        provider:
          type: string
        type:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        author:
          type: ["string", "null"]
        artist:
          type: ["string", "null"]
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/Chapter"
        totalChapters:
          type: integer
        alternativeTitles:
          type: array
          items:
            type: string
        serialization:
          type: ["string", "null"]
        updatedAt:
          type: ["string", "null"]
      required:
        - id
        - title
        - coverImage
        - provider
        - genres
        - chapters
        - totalChapters
        - alternativeTitles
    ChapterPages:
      type: object
      properties:
        seriesId:
          type: string
        chapterId:
          type: string
        provider:
          type: string
        pages:
          type: array
          items:
            type: string
        currentChapter:
          type: string
        prevChapterId:
          type: ["string", "null"]
        nextChapterId:
          type: ["string", "null"]
        seriesTitle:
          type: string
        chapterTitle:
          type: ["string", "null"]
      required:
        - seriesId
        - chapterId
        - provider
        - pages
        - currentChapter
        - seriesTitle
...
    FavouriteItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        provider:
          type: string
        seriesId:
          type: string
        title:
          type: string
        coverImage:
          type: string
        type:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - provider
        - seriesId
        - title
        - coverImage
        - createdAt
    FavouriteInput:
      type: object
      properties:
        provider:
          type: string
        seriesId:
          type: string
        title:
          type: string
        coverImage:
          type: string
        type:
          type: string
        status:
          type: string
      required:
        - provider
        - seriesId
        - title
        - coverImage
...