# lib/api-zod/src/generated/types/favouriteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShiroScans API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FavouriteInput {
  provider: string;
  seriesId: string;
  title: string;
  coverImage: string;
  type?: string;
  status?: string;
}